Add the basic template for book-appointment page
The book-appointment page shows all the available appointment time slots that a patient can book. The entire functionality of this page is still not completed, styles applied to this page is temporay and will change in the near future. A static directory has been created with sub-directories for the css, scripts and any images that has been used in the book-appointment page. CSS has been aded to the html elements in an inline fashion, styles will be seperated out to external style sheet in the near future.

The book-appointment-details page is a page to display a detailed view of the selected appointment time slot, (This page may not be used if the structure of the interaction changes).