Add Date and Time picker to Practitioners set-appointments page.
The required JQuery Library files to make the date and time pickeris included.
The maximum DoB has been increased to last 100 years (Patient Sign-up page).
Add a 'Back' button to the view that shows when an item is deleted from the cart.
Add some styling to that view.

This resloves issue #120