Fix all Design related bugs and errors that were found
-(base)
Renamed Sign-up to Register in base navigation bar.

-(Patient)
Fix a minor bug in patient profile page.
Change input type to submit for reset password button.

-(Bookings)
Add a grid system to present the available bookable slots in a neat way.
Add responsive style for merge info-alert.
Change username to get_full_name.

-(Practitioner)
Fix footer to be relative to the content, as the content in that page doesn't grow.